src/pyproject.toml
[tool.poetry]
name = "TgVoiceMailer"
version = "0.1.0"
description = ""
authors = ["Yegor Shevchenko <egorhs19@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
gcloud = "^0.18.3"
google-cloud-storage = "^1.44.0"
python-telegram-bot = "^13.10"
betterconf = "^2.6.1"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
types-requests = "^2.27.8"

[tool.poetry.dev-dependencies]
pydocstyle = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
